# apps/api/openapi.json
{
  "components": {
    "schemas": {
...
      "FreshnessPolicyView": {
        "additionalProperties": false,
        "description": "Coordonnées SERVEUR de la jauge âge / budget d'une réponse datable.\n\n``budget_seconds`` est le TTL de SÉANCE FERMÉE de la politique du registre\n(`vertex_core.data.freshness`) qui borne le relais — la même valeur, du\nmême propriétaire (`vertex_api.freshness.closed_session_budget`), que\ncelle nommée dans la raison ``stale``. ``kind`` est le NOM de cette\npolitique, c'est-à-dire la famille d'observation la plus fraîche dont\nl'instantané peut être issu (``daily_bar``, ``news_attention``,\n``option_surface``…) — à ne pas confondre avec le ``kind`` d'un\ninstantané. ``version`` est la version de la politique : tout changement\nde TTL exige une montée de version, la jauge la porte donc avec elle.\n\nLe client pose ``age_seconds`` sur cette échelle et n'invente ni TTL ni\nratio : publier le budget évite un second registre recopié côté\ninterface. Le budget est une propriété de la ROUTE, pas de l'instantané :\nil est servi dans tous les états, ``empty`` compris — sans instantané\nl'âge est ``null`` (il n'existe pas), l'échelle reste connue. Un budget\nnul est refusé à la frontière : c'est la forme qu'une absence prendrait\nsi elle était convertie en zéro. Une famille sans budget au registre\npublie ``null`` (matrice de capacités), jamais un TTL inventé.",
        "properties": {
          "budget_seconds": {
            "exclusiveMinimum": 0.0,
            "title": "Budget Seconds",
            "type": "integer"
          },
          "kind": {
            "minLength": 1,
            "title": "Kind",
            "type": "string"
          },
          "version": {
            "minLength": 1,
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "budget_seconds",
          "kind",
          "version"
        ],
        "title": "FreshnessPolicyView",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OptionChainContract": {
        "additionalProperties": false,
        "description": "One option contract row, relayed from the worker snapshot verbatim.\n\nThe COMPLETE contract identity (synthetic ``con_id``, trading class,\nstrike, right, multiplier, currency, exchange, style, settlement,\nexpiration) travels with every row. ``quote`` is the verbatim observed\nquote plus its quality status (``OK``/``CROSSED``/``STALE``/``MISSING``);\n``iv`` and ``greeks`` are the worker's Vertex results — present ONLY when\nthe quote was sane, labeled ``value_nature = \"THEORETICAL\"`` with their\npreserved ``CalculationRecord`` lineage, otherwise honestly ``ABSENT``\nwith the typed refusal reason. The API recomputes nothing.",
        "properties": {
          "con_id": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Con Id"
          },
          "currency": {
            "minLength": 1,
            "title": "Currency",
            "type": "string"
          },
          "exchange": {
            "minLength": 1,
            "title": "Exchange",
            "type": "string"
          },
          "expiration": {
            "minLength": 1,
            "title": "Expiration",
            "type": "string"
          },
          "greeks": {
            "additionalProperties": true,
            "title": "Greeks",
            "type": "object"
          },
          "iv": {
            "additionalProperties": true,
            "title": "Iv",
            "type": "object"
          },
          "multiplier": {
            "exclusiveMinimum": 0.0,
            "title": "Multiplier",
            "type": "integer"
          },
          "open_interest": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Open Interest"
          },
          "open_interest_status": {
            "anyOf": [
              {
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Open Interest Status"
          },
          "quote": {
            "additionalProperties": true,
            "title": "Quote",
            "type": "object"
          },
          "right": {
            "anyOf": [
              {
                "enum": [
                  "CALL",
                  "PUT"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Right"
          },
          "settlement": {
            "minLength": 1,
            "title": "Settlement",
            "type": "string"
          },
          "strike": {
            "anyOf": [
              {
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Strike"
          },
          "style": {
            "minLength": 1,
            "title": "Style",
            "type": "string"
          },
          "synthetic": {
            "title": "Synthetic",
            "type": "boolean"
          },
          "trading_class": {
            "minLength": 1,
            "title": "Trading Class",
            "type": "string"
          },
          "volume": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Volume"
          }
        },
        "required": [
          "con_id",
          "strike",
          "right",
          "expiration",
          "trading_class",
          "multiplier",
          "currency",
          "exchange",
          "style",
          "settlement",
          "quote",
          "volume",
          "open_interest",
          "open_interest_status",
          "iv",
          "greeks",
          "synthetic"
        ],
        "title": "OptionChainContract",
        "type": "object"
      },
      "OptionChainExpiration": {
        "additionalProperties": false,
        "description": "One ``(expiration, trading_class)`` group — never merged by date.\n\nTwo trading classes at the same expiration date are two distinct groups\n(distinct identities). ``coverage`` is the worker's honest account:\nexpected/received/valid/resolved contracts and the per-contract discard\nreasons.",
        "properties": {
          "contracts": {
            "items": {
              "$ref": "#/components/schemas/OptionChainContract"
            },
            "title": "Contracts",
            "type": "array"
          },
          "coverage": {
            "additionalProperties": true,
            "title": "Coverage",
            "type": "object"
          },
          "currency": {
            "minLength": 1,
            "title": "Currency",
            "type": "string"
          },
          "exchange": {
            "minLength": 1,
            "title": "Exchange",
            "type": "string"
          },
          "expiration": {
            "minLength": 1,
            "title": "Expiration",
            "type": "string"
          },
          "maturity_years": {
            "minLength": 1,
            "title": "Maturity Years",
            "type": "string"
          },
          "multiplier": {
            "exclusiveMinimum": 0.0,
            "title": "Multiplier",
            "type": "integer"
          },
          "quality": {
            "minLength": 1,
            "title": "Quality",
            "type": "string"
          },
          "settlement": {
            "minLength": 1,
            "title": "Settlement",
            "type": "string"
          },
          "source_event_id": {
            "minLength": 1,
            "title": "Source Event Id",
            "type": "string"
          },
          "style": {
            "minLength": 1,
            "title": "Style",
            "type": "string"
          },
          "trading_class": {
            "minLength": 1,
            "title": "Trading Class",
            "type": "string"
          }
        },
        "required": [
          "expiration",
          "trading_class",
          "exchange",
          "style",
          "settlement",
          "multiplier",
          "currency",
          "maturity_years",
          "quality",
          "source_event_id",
          "contracts",
          "coverage"
        ],
        "title": "OptionChainExpiration",
        "type": "object"
      },
      "OptionChainResponse": {
        "additionalProperties": false,
        "description": "The last ``option_chain/{underlying}`` snapshot — or an honest empty.\n\n``state = \"empty\"`` means NO snapshot was ever published for this\nunderlying: every snapshot-derived field is ``None`` (never invented) and\n``reason`` says why. ``state = \"ok\"`` relays the persisted content\nverbatim: population (``REAL``/``SYNTHETIC`` shown as-is), the published\nspot, the pricing assumptions, the per-(expiration, trading_class) groups\nand the displayed row budget.\n\n``state = \"stale\"`` relaie le MÊME contenu, mais dit que l'instantané\na dépassé son budget de fraîcheur (option_surface) : le worker n'a rien\npublié de plus récent. ``age_seconds`` est publié dans TOUS les états\ndatables — son absence faisait passer un instantané de trois jours\npour un instantané d'une minute. ``freshness_policy`` publie le budget\ncontre lequel cet âge est jugé (``FreshnessPolicyView``).\n\n``spot`` relaie le bloc du worker VERBATIM : ``value``, ``currency``,\npuis la provenance PROPRE du spot — ``basis`` (``daily_close`` pour le\ncollecteur réel), ``observed_at`` (l'instant DU SPOT, jamais celui de\nla tranche d'options), ``source_event_id`` (l'observation qui l'a\nfourni), ``carried_by_event_id`` (la tranche qui l'a porté),\n``provenance`` (``PUBLISHED`` | ``PARTIAL`` | ``NOT_PUBLISHED``),\n``age_seconds``, ``max_age_seconds`` (borne DÉCLARÉE par le worker) et\n``age_status`` (``OK`` | ``STALE`` | ``FUTURE`` | ``UNKNOWN``). Un\nchamp absent est ``null`` et le dit ; l'API ne le comble pas.",
        "properties": {
          "age_seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Age Seconds"
          },
          "as_of": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "As Of"
          },
          "assumptions": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assumptions"
          },
          "coverage": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Coverage"
          },
          "engine_version": {
            "anyOf": [
              {
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Engine Version"
          },
          "expirations": {
            "items": {
              "$ref": "#/components/schemas/OptionChainExpiration"
            },
            "title": "Expirations",
            "type": "array"
          },
          "freshness_policy": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FreshnessPolicyView"
              },
              {
                "type": "null"
              }
            ]
          },
          "population": {
            "anyOf": [
              {
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Population"
          },
          "reason": {
            "anyOf": [
              {
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "row_budget": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Row Budget"
          },
          "snapshot_version": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Snapshot Version"
          },
          "spot": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Spot"
          },
          "state": {
            "enum": [
              "ok",
              "stale",
              "empty"
            ],
            "title": "State",
            "type": "string"
          },
          "underlying": {
            "minLength": 1,
            "title": "Underlying",
            "type": "string"
          },
          "value_nature": {
            "anyOf": [
              {
                "const": "THEORETICAL",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value Nature"
          }
        },
        "required": [
          "state",
          "snapshot_version",
          "as_of",
          "age_seconds",
          "freshness_policy",
          "population",
          "underlying",
          "engine_version",
          "value_nature",
          "spot",
          "assumptions",
          "expirations",
          "row_budget",
          "coverage",
          "reason"
        ],
        "title": "OptionChainResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Local, analysis-only API of Vertex One. It serves the canonical contracts of vertex_core and delegates every verdict to the single AdviceEngine. No transactional capability exists; the human remains the sole decision maker, acting outside Vertex.",
    "title": "Vertex One API",
    "version": "vertex_core@0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/options/{underlying}/chain": {
      "get": {
        "description": "Serve the LAST ``option_chain/{underlying}`` snapshot exactly as persisted.\n\nThe API relays the worker's published content — per-(expiration,\ntrading_class) groups with complete contract identities, verbatim quotes\nand their quality, the worker's Vertex IV/Greeks (``THEORETICAL``, with\ntheir ``CalculationRecord`` lineage) or their typed refusal reasons, the\ncoverage account and the displayed row budget — and computes nothing.\nWith no snapshot ever published for this underlying the answer is a 200\nwith ``state = \"empty\"``: absent stays absent, nothing is invented.",
        "operationId": "get_option_chain",
        "parameters": [
          {
            "in": "path",
            "name": "underlying",
            "required": true,
            "schema": {
              "pattern": "^[A-Za-z0-9][A-Za-z0-9._-]{0,63}$",
              "title": "Underlying",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OptionChainResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "Authentication required: no valid WebAuthn session cookie (or missing/invalid CSRF header on a mutation). Always the same generic body with detail code AUTH_REQUIRED."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Last published option chain for one underlying (or honest empty state)"
      }
    },
...
  }
}